move view_vectors in helpers and fix circular import
